@@ -19,8 +19,8 @@
     "moment": "^2.9.0",
     "mousetrap": "^1.4.6",
     "oboe": "^2.1.2",
-    "react": "^0.12.2",
-    "react-tools": "^0.12.2",
+    "react": "^0.13.2",
+    "react-tools": "^0.13.2",
     "underscore": "^1.7.0"
   },
   "devDependencies": {
